update Back-end
- Update "user.model.js" with every data we'll be needing
- Create "readme.md" with instructions on how to up the node.js server
- Using English for the everything
