+++ App/backend-casino/.env
@@ -1,6 +1,6 @@
 PORT = 3000
 DB_USER = 'root'
-DB_PASSWORD = '1234'
+DB_PASSWORD = 'supergg2'
 DB_HOST = 'localhost'
 DB_NAME = 'casino_test'
 DB_PORT = '3306'
