+++ App/backend-casino/readme.md
@@ -1,0 +1,23 @@
+[1] npm i
+
+[2] Poner los datos de nuestra conexion a la base de datos en ".env" <br>
+Lo mas probable es que se necesite cambiar DB_USER, DB_PASSWORD, DB_HOST
+
+[3] Crear una base de datos llamada "casino_test" si no esta creada antes
+
+[4] npm run dev <br>
+Para inicializar el servidor y crear las tablas correspondientes. 
+
+### Comandos de Postman
+#### PUT - Users
+
+    "first_name": "",
+    "last_name": "",
+    "birthday": "YYYY/DD/MM",
+    "street": "",
+    "phone": "",
+    "email": "",
+    "password": "",
+    "role": "",
+    "balance": ""
+
