+++ App/backend-casino/src/db/models/index.js
@@ -1,4 +1,4 @@
-const { User, UserSchema } = require('./usuario.model');
+const { User, UserSchema } = require('./user.model');
 
 function setupModels(sequelize){
     User.init(UserSchema, User.config(sequelize));
